+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
+## v1.3.0
+
+- add clientset creator
+- add alerts contains
+
 ## v1.2.0
 
 - add AlertDeployer